requirements.txt
streamlit
streamlit-extras
streamlit-code-editor
streamlit-mic-recorder
streamlit-TTS
streamlit_ace
matplotlib
numpy
pandas
graphviz
altair
plotly
bokeh
pydeck
scipy
sympy
scikit-learn
vega-datasets
python-socketio
eventlet
firebase_admin
openai
requests
tiktoken
seleniumbase
get-gecko-driver
beautifulsoup4
PyPDF2
python-docx
odfpy
typing_extensions
pygithub
streamlit-ace
streamlit_code_editor